separate featurecard from entercoorswindow
